refactor: move services to a new services folder
