+++ src/app/[soorah]/page.tsx
@@ -5,9 +5,8 @@ import { sirasayi } from 'sirasayi'
 import { Bismillah } from '@/app/[soorah]/components/Bismillah'
 import { soorahList } from '@/assets/soorah-list-object'
 import { TemplateAyahList } from '@/components/TemplateAyahList'
+import { getSoorahService } from '@/servises/getSoorahService'
 import { getView } from '@/utility/getView/getView'
-
-import { getSoorahService } from './getSoorahService'
 
 type SoorahPageProps = {
   params: Promise<{
